+++ src/main/java/com/group3/evproject/repository/ChargingSessionRepository.java
@@ -5,14 +5,14 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ChargingSessionRepository extends  JpaRepository<ChargingSession,Integer> {
-    List<ChargingSession> findByBooking_Id(Integer bookingId);
+public interface ChargingSessionRepository extends  JpaRepository<ChargingSession,Long> {
+    List<ChargingSession> findByBooking_Id(Long bookingId);
 
-    List<ChargingSession> findByStation_Id(Integer stationId);
+    List<ChargingSession> findByStation_Id(Long stationId);
 
     List<ChargingSession> findByStatus(String status);
 
-    List<ChargingSession> findBySpot_Id(Integer spotId);
+    List<ChargingSession> findBySpot_Id(Long spotId);
 
     List<ChargingSession> findByStartTimeBetween(java.time.LocalDateTime start, java.time.LocalDateTime end);
 }
